Revert "fstests: udev settle before _scratch_pool_mkfs"
This reverts commit 6c655399e08a6dd05a4909e600efe66da523025f.

It's quite possible that this patch to fstests is not necessary

https://lore.kernel.org/fstests/30c274a6-adc1-42a7-91cf-8a386884106f@oracle.com/

points to patch e54514aaeab6be ("btrfs-progs: fix stray fd close in
open_ctree_fs_info()") in btrfs-progs that fixes handling of file
descriptors so udev and mkfs work together.

This is from btrfs-progs 6.7.1, older version might still fail but we're
now gathering data.

Signed-off-by: David Sterba <dsterba@suse.com>